backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
joblib
anthropic
python-dotenv